+++ Drops.cs
@@ -1,6 +1,6 @@
 ﻿using MultidimensionMod.Items.Accessories;
 using MultidimensionMod.Items.Materials;
-using MultidimensionMod.Items.Weapons.Meele.Broadswords;
+//using MultidimensionMod.Items.Weapons.Meele.Broadswords;
 using MultidimensionMod.Items.Weapons.Magic.Guns;
 using MultidimensionMod.Items.Weapons.Ranged;
 using Terraria;
