+++ Items/Weapons/Magic/Staffs/StormScepter.cs
@@ -26,7 +26,7 @@ namespace MultidimensionMod.Items.Weapons.Magic.Staffs
 
 		public override bool Shoot(Player player, ref Vector2 position, ref float speedX, ref float speedY, ref int type, ref int damage, ref float knockBack)
 		{
-			type = ModContent.ProjectileType<StormDroplet>();
+			//type = ModContent.ProjectileType<StormDroplet>();
 			return base.Shoot(player, ref position, ref speedX, ref speedY, ref type, ref damage, ref knockBack);
 		}
 	}
