+++ MultidimensionMod.cs
@@ -1,7 +1,7 @@
 using MultidimensionMod.Items.Accessories;
 using MultidimensionMod.Items.Materials;
-using MultidimensionMod.Items.Weapons.Meele.Boomerangs;
-using MultidimensionMod.Items.Weapons.Meele.Broadswords;
+//using MultidimensionMod.Items.Weapons.Meele.Boomerangs;
+//using MultidimensionMod.Items.Weapons.Meele.Broadswords;
 using MultidimensionMod.Items.Weapons.Magic.Guns;
 using MultidimensionMod.Items.Weapons.Ranged;
 using System;
@@ -121,7 +121,7 @@ namespace MultidimensionMod
 				{
 					ModContent.ItemType<EyeTendril>(),
 					ModContent.ItemType<EyeoftheNightwalker>(),
-					ModContent.ItemType<Iris>()
+					//ModContent.ItemType<Iris>()
 				});
 				MultidimensionMod.AddLoot(bossChecklist, "SkeletronHead", new List<int>
 				{
@@ -135,7 +135,7 @@ namespace MultidimensionMod
 				MultidimensionMod.AddLoot(bossChecklist, "TheTwins", new List<int>
 				{
 					ModContent.ItemType<Retilazor>(),
-					ModContent.ItemType<Spazmelter>()
+					//ModContent.ItemType<Spazmelter>()
 				});
 				MultidimensionMod.AddLoot(bossChecklist, "Plantera", new List<int>
 				{
@@ -144,11 +144,11 @@ namespace MultidimensionMod
 				MultidimensionMod.AddLoot(bossChecklist, "DukeFishron", new List<int>
 				{
 					ModContent.ItemType<TidalQuartz>(),
-					ModContent.ItemType<TyphoonDragon>()
+					//ModContent.ItemType<TyphoonDragon>()
 				});
 				MultidimensionMod.AddLoot(bossChecklist, "Solar Eclipse", new List<int>
 				{
-					ModContent.ItemType<EclipseReaper>()
+					//ModContent.ItemType<EclipseReaper>()
 				});
 				MultidimensionMod.AddLoot(bossChecklist, "Frost Moon", new List<int>
 				{
