+++ Projectiles/Boss/Smiley/Darkling.cs
@@ -7,14 +7,12 @@ using Terraria.ModLoader;
 
 namespace MultidimensionMod.Projectiles.Boss.Smiley
 {
-
-    public class Baller2 : ModProjectile
+    public class Darkling : ModProjectile
     {
-        private int bounces;
         public override void SetStaticDefaults()
         {
             DisplayName.SetDefault("SMILE!");
-            Main.projFrames[projectile.type] = 1; //This is an animated projectile
+            Main.projFrames[projectile.type] = 4; //This is an animated projectile
 
         }
         public override void SetDefaults()
@@ -25,7 +23,7 @@ namespace MultidimensionMod.Projectiles.Boss.Smiley
             projectile.penetrate = 5;                       //this is the projectile penetration
             projectile.hostile = true;
             projectile.magic = true;                        //this make the projectile do magic damage
-            projectile.tileCollide = true;                 //this make that the projectile does not go thru walls
+            projectile.tileCollide = false;                 //this make that the projectile does not go thru walls
             projectile.ignoreWater = false;
             projectile.light = 0.4f;    // projectile light
 
@@ -44,25 +42,6 @@ namespace MultidimensionMod.Projectiles.Boss.Smiley
             }
 
         }
-        public override bool OnTileCollide(Vector2 oldVelocity)
-        {
-            if (projectile.velocity.X != oldVelocity.X)
-            {
-                projectile.position.X = projectile.position.X + projectile.velocity.X;
-                projectile.velocity.X = -oldVelocity.X;
-            }
-            if (projectile.velocity.Y != oldVelocity.Y)
-            {
-                projectile.position.Y = projectile.position.Y + projectile.velocity.Y;
-                projectile.velocity.Y = -oldVelocity.Y;
-            }
-            bounces++;
-            if(bounces >6)
-            {
-                projectile.Kill();
-            }
-            return false; // return false because we are handling collision
-        }
         public override void OnHitPlayer(Player target, int damage, bool crit)
         {
             target.AddBuff(BuffID.OnFire, 150);
@@ -71,13 +50,11 @@ namespace MultidimensionMod.Projectiles.Boss.Smiley
 
         public override void AI()
         {
+            projectile.velocity *= 1.01f;
             projectile.spriteDirection = -1 * projectile.direction;
 
             //this is projectile dust
-            if (projectile.tileCollide)
-            {
-             //   projectile.velocity *= -1;
-            }
+
             int DustID2 = Dust.NewDust(projectile.position + projectile.velocity, projectile.width, projectile.height, DustID.DungeonSpirit, projectile.velocity.X * -0.1f, projectile.velocity.Y * -0.1f, 0, default, 1.25f);   //spawns dust behind it, this is a spectral light blue dust lol
             Main.dust[DustID2].noGravity = true;
 
@@ -85,20 +62,19 @@ namespace MultidimensionMod.Projectiles.Boss.Smiley
             //this make that the projectile faces the right way
             projectile.rotation = (float)Math.Atan2((double)projectile.velocity.Y, (double)projectile.velocity.X) + 1.57f;
 
-
+            projectile.timeLeft = 999;
             projectile.localAI[0] += 1f;
             //projectile.alpha = (int)projectile.localAI[0] * 2;
 
-            if (projectile.localAI[0] > 5 * 60f) //projectile time left before disappears
-            {
-
-                projectile.Kill();
-            }
+            
             // Loop through the 20 animation frames, spending 15 ticks on each.
-            if (++projectile.frameCounter >= 30)
+            int frameSpeed = 8;
+            projectile.frameCounter++;
+            if (projectile.frameCounter >= frameSpeed)
             {
                 projectile.frameCounter = 0;
-                if (++projectile.frame >= 20)
+                projectile.frame++;
+                if (projectile.frame >= Main.projFrames[projectile.type])
                 {
                     projectile.frame = 0;
                 }
